requirements.txt
python-telegram-bot==20.3
supabase==1.0.3
python-dotenv
